imgui: switch to upstream dear_bindings
Manually convert pointers to references.

Change const& variables to values (needed because) the convert
references to pointers pass was also doing this conversion but now we
are not running this pass anymore.
This is not ideal but is quite hard to change because then the generator
is not generating some of the methods because those const refs gets
transformed into pointers later. The overhead of the value to const ref
is likely negligible for us.

We can think about this later if it ever becomes an issue.